+++ jaguar/lib/src/serve/server.dart
@@ -51,10 +51,14 @@ class Jaguar extends Object with Muxable {
   /// Defaults to [JaguarSessionManager].
   SessionManager sessionManager;
 
+  final before = <RouteInterceptor>[];
+
+  final after = <RouteInterceptor>[];
+
   /// Logger used to log concise useful information about the request. This is
   /// also available in [Context] so that interceptors and route handlers can also
   /// log.
-  Logger log = new Logger('J');
+  final log = Logger('J');
 
   /// Returns protocol string
   String get protocolStr => securityContext == null ? 'http' : 'https';
@@ -116,9 +120,11 @@ class Jaguar extends Object with Muxable {
       });
 
   Future<void> _handler(HttpRequest request) async {
-    final ctx = new Context(new Request(request),
+    final ctx = Context(Request(request),
         sessionManager: sessionManager, log: log, userFetchers: userFetchers);
     ctx.prefix = basePath;
+    ctx.before.addAll(before);
+    ctx.after.addAll(after);
 
     try {
       // Try to find a matching route and invoke it.
@@ -128,16 +134,13 @@ class Jaguar extends Object with Muxable {
       }
 
       // If no response, write 404 error.
-      if (ctx.response == null) {
-        errorWriter.make404(ctx);
-      }
+      if (ctx.response == null) errorWriter.make404(ctx);
     } catch (e, stack) {
       if (e is Response) {
         // If [Response] object was thrown, write it!
         ctx.response = e;
-      } else {
+      } else
         errorWriter.make500(ctx, e, stack);
-      }
     }
 
     try {
@@ -180,9 +183,8 @@ class Jaguar extends Object with Muxable {
 
   /// Adds the [Route] to be served
   Route addRoute(Route route) {
-    if (_server != null) {
-      throw new Exception('Cannot add routes after server has been started!');
-    }
+    if (_server != null)
+      throw Exception('Cannot add routes after server has been started!');
     _handlers.add(route);
     return route;
   }